requirements.txt
pillow
lxml
pycrypto
greenlet
protobuf-py3
asyncio
statsd
asyncio_redis
websockets
PyYAML
coloredlogs
